requirements.txt
cryptography<39,>=2.8
chess<2,>=1
pyzmq<25,>=19
